src/requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==4.0.0
django-filter==23.2
django-jazzmin==2.6.0
django-ratelimit2==0.4.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2022.7.1
requests==2.30.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
telegram==0.0.1
tomli==2.0.1
twilio==7.17.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.2
webencodings==0.5.1
